+++ heroku/utils/args.py
@@ -86,7 +86,7 @@ def get_args_html(message: Message) -> str:
     :param message: Message to get arguments from
     :return: String with HTML arguments
     """
-    prefix = message.client.loader.get_prefix()
+    prefix = message._client.loader.get_prefix()
 
     if not (message := message.text.html):
         return False